Move GPU-related Hcal records to CondFormats/DataRecord
