+++ CondFormats/DataRecord/src/HcalCombinedRecordsGPU.cc
@@ -1,5 +1,5 @@
+#include "CondFormats/DataRecord/interface/HcalCombinedRecordsGPU.h"
 #include "FWCore/Framework/interface/eventsetuprecord_registration_macro.h"
-#include "CondFormats/HcalObjects/interface/HcalCombinedRecordsGPU.h"
 
 EVENTSETUP_RECORD_REG(HcalConvertedPedestalsRcd);
 EVENTSETUP_RECORD_REG(HcalConvertedPedestalWidthsRcd);
