+++ CondFormats/HcalObjects/BuildFile.xml
@@ -1,16 +1,15 @@
 <ifrelease name="_UBSAN_">
   <flags REM_CXXFLAGS="-fno-omit-frame-pointer -fsanitize=undefined"/>
 </ifrelease>
+<use name="boost_serialization"/>
+<use name="CondFormats/Serialization"/>
 <use name="DataFormats/DetId"/>
 <use name="DataFormats/HcalDetId"/>
-<use name="Geometry/CaloTopology"/>
-<use name="FWCore/Utilities"/>
 <use name="FWCore/MessageLogger"/>
-<use name="CondFormats/Serialization"/>
+<use name="FWCore/Utilities"/>
+<use name="Geometry/CaloTopology"/>
 <use name="HeterogeneousCore/CUDACore"/>
 <use name="HeterogeneousCore/CUDAUtilities"/>
-<use name="CondFormats/DataRecord"/>
-<use name="boost_serialization"/>
 <export>
   <lib name="1"/>
 </export>
